+++ tests/helpers/__init__.py
@@ -2,16 +2,17 @@
 #  Permissions are hereby granted under the terms of the Apache 2.0 License:
 #  https://opensource.org/license/apache-2-0.
 
-from .sentinel1 import make_s1_grd_datatree
+from .sentinel1 import make_s1_grd_datatree, make_s1_ocn_datatree
 from .sentinel2 import make_s2_msi, make_s2_msi_l1c, make_s2_msi_l2a
 from .sentinel3 import make_s3_olci_efr, make_s3_slstr_lst, make_s3_slstr_rbt
 
 __all__ = [
+    "make_s1_grd_datatree",
+    "make_s1_ocn_datatree",
     "make_s2_msi",
     "make_s2_msi_l1c",
     "make_s2_msi_l2a",
     "make_s3_olci_efr",
     "make_s3_slstr_rbt",
     "make_s3_slstr_lst",
-    "make_s1_grd_datatree",
 ]